Add nvim and tmux config files
It's been a few years but after using VSCode for a bit, I thought it
would be a great time to try and learn `vim/nvim` again. To supplement
this, I'm also learning `tmux` (again) to round off my productivity and
dev toolkit. I've been using these tools both for personal projects and
at work for almost a week now, but I'm going to try and stick with it
for a couple of months so I can get a good feel and hopefully become
more productive with them.